Restrict contact access to authenticated users and enhance authorization checks

+++ controllers/contactController.js
@@ -6,7 +6,7 @@ const Contact = require("../models/contactModel");
 //@route GET /api/contacts
 //@access private
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({user_id: req.user.id});
   res.status(200).json(contacts);
 });
 
@@ -25,6 +25,7 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id
   });
 
   res.status(201).json(contact);
@@ -52,6 +53,11 @@ const updateContact = asyncHandler(async (req, res) => {
     throw new Error("Contact not found");
   }
 
+  if(contact.user_id.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("Not authorized to update this contact");
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -75,6 +81,11 @@ const deleteContact = asyncHandler(async (req, res) => {
       console.log("Contact not found");
       res.status(404);
       throw new Error("Contact not found");
+    }
+
+    if(contact.user_id.toString() !== req.user.id) {
+      res.status(401);
+      throw new Error("Not authorized to update this contact");
     }
 
     await contact.deleteOne();
